Add files via upload
Source file for Fake News Detection model using the the model I've chosen to use is the Passive-Aggressive (PA) Classifier (see original paper here). In essence, the PA classifier is an algorithm that only updates its weights ("aggressive" action) when it encounters examples for which its predictions are wrong, but otherwise remains unchanged ("passive" action).

The PA classifier is an online algorithm, meaning it uses one example at a time to update its weights and moves on, never seeing the same example again. This is in contrast to a batch algorithm, which would use the same set of multiple examples and updates weights in each iteration of training. Because of this, the PA classifier is particularly useful when dealing with a dataset containing a large or rapidly increasing number of examples, like news articles or Tweets! Of course, the data I'm using in this notebook are toy static data, but you can imagine its advantages in real-life applications. Other Kagglers, like Ayushi Mishra have shown that the PA classifier outperforms several other types of models as well, so I can be confident that it is a good choice.

If you'd like to learn more about the mathematics behind the PA classifier algorithm, check out this video by Dr. Victor Lavrenko that explains the steps in very clear steps!